requirements.txt
pydantic~=1.10.4
sentry-sdk~=1.25.0
fastapi~=0.97.0
uvicorn~=0.20.0
pytest~=7.3.2
mlflow
hydra-core
torch
pytorch-lightning
numpy
networkx
transformers
torchmetrics
dvc[s3]
joblib
sacremoses
textdistance
fastapi-jwt-auth
../ai-es-utils
